requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.0
python-decouple==3.8
djangorestframework==3.14.0
django-filter==23.3
requests==2.31.0